requirements.txt
spotipy
numpy
python-dotenv